requirements.txt
streamlit
pandas
taos-ws-py
openai
python-dotenv
matplotlib
chardet
loguru
PyYAML
scikit-learn
